Use double instead of float for generated types
Qt uses double mostly. In QML all reals are stored in double precision.
DBus also works with doubles.
Clients also use double more often than float. So it would be better if
generator will work with double instead of float to avoid conversion
from double to float and prevent precision lost.

+++ src/desktop-dev-tools/DummyModel.h
@@ -92,7 +92,7 @@ inline void readJSONSimple<bool>(const QJsonValue &json, bool &value)
 }
 
 template<>
-inline void readJSONSimple<float>(const QJsonValue &json, float &value)
+inline void readJSONSimple<double>(const QJsonValue &json, double &value)
 {
     if (json.isDouble()) {
         value = json.toDouble();
